+++ R/MAINest.R
@@ -191,7 +191,7 @@
 #' @export
 
 fitGMVAR <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "mean"), constraints=NULL, structural_pars=NULL,
-                     ncalls=round(10 + 9*log(M)), ncores=min(2, ncalls, parallel::detectCores()), maxit=300, seeds=NULL,
+                     ncalls=round(10 + 9*log(M)), ncores=min(2, ncalls, parallel::detectCores()), maxit=500, seeds=NULL,
                      print_res=TRUE, ...) {
 
   on.exit(closeAllConnections())